requirements.txt
python-telegram-bot
schedule
requests
asyncio
os
